requirements.txt
pip~=22.3.1
requests~=2.28.1
Django~=4.1.5
djangorestframework~=3.14.0
django-unixdatetimefield==1.2.0
django-cors-headers~=3.13.0
Faker~=16.1.0